+++ drizzle/20260318043126_restricted/migration.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `botData` ADD `restricted` integer DEFAULT false;